Hemant

+++ FoodERP/FoodERPApp/Views/V_PurchaseReturn.py
@@ -290,10 +290,21 @@ class PurchaseReturnView(CreateAPIView):
                         PurchaseReturn_Data.delete()        
                         return JsonResponse({'StatusCode': 200, 'Status': True, 'Message': 'Return Deleted Successfully', 'Data': []}) 
                 else:
-                    PurchaseReturn_Data = T_PurchaseReturn.objects.get(id=id)
-                    PurchaseReturn_Data.delete()
-                    return JsonResponse({'StatusCode': 200, 'Status': True, 'Message': 'Return Deleted Successfully', 'Data': []})
-                    
+                    Query = T_PurchaseReturn.objects.filter(id=id)
+                    if Query.exists():
+                        PurchaseReturnSerializer = PurchaseReturnSerializerThird(Query, many=True).data
+                        for a in PurchaseReturnSerializer:
+                            for b in a['ReturnItems']:
+                                OBatchQuantity=O_BatchWiseLiveStock.objects.filter(id=b['BatchID'],Item=b['Item']['id'],Unit=b['Unit']['id']).values('OriginalBaseUnitQuantity','BaseUnitQuantity')
+                                Qty=float(OBatchQuantity[0]['BaseUnitQuantity']) + float(b['BaseUnitQuantity'])
+                                if(OBatchQuantity[0]['OriginalBaseUnitQuantity'] <= float(Qty)):
+                                    OBatchWiseLiveStock=O_BatchWiseLiveStock.objects.filter(PurchaseReturn=b['SubReturn'],Item=b['Item']['id']).update(BaseUnitQuantity = Qty ) 
+                                    Qty =0.00
+                                else:    
+                                    return JsonResponse({'StatusCode': 200, 'Status': True, 'Message': 'Return Qty greater than Consolidated return qty', 'Data': []})
+                        PurchaseReturn_Data = T_PurchaseReturn.objects.get(id=id)
+                        PurchaseReturn_Data.delete()    
+                        return JsonResponse({'StatusCode': 200, 'Status': True, 'Message': 'Return Deleted Successfully', 'Data': []})    
         except T_PurchaseReturn.DoesNotExist:
             return JsonResponse({'StatusCode': 204, 'Status': True, 'Message': 'Record Not available', 'Data': []})
         except IntegrityError:
